+++ ui/sweep.h
@@ -11,7 +11,7 @@ namespace UI {
 // radius: radius of tube
 // vertex: vertices for the tube
 // triangles: triangles for the tube
-void circleSweep(const Eigen::MatrixXf path, const float radius, const int stride,
-    Eigen::MatrixXf *vertex, Eigen::MatrixXi *triangles);
+void circleSweep(const Eigen::MatrixXd path, const double radius, const int stride,
+    Eigen::MatrixXd *vertex, Eigen::MatrixXi *triangles);
 
 }  // namespace UI
